requirements/base.txt
Django==2.2
psycopg2>=2.8.1
pillow>=5.0.0
djangorestframework>=3.6.2
django_filter>=1.0.2
django-simple-history>=1.8.2
coreapi>=2.3.0
pygments>=2.2.0
django-crispy-forms>=1.7.2
django-import-export>=1.0.0
django-cleanup>=2.1.0
django-qr-code==1.0.0
